added readme and renamed
