+++ gomonorepo/c_lint_modules.go
@@ -1,4 +1,4 @@
-package gomono
+package gomonorepo
 
 import (
 	"context"
