+++ gomonorepo/c_test_modules_test.go
@@ -1,4 +1,4 @@
-package gomono
+package gomonorepo
 
 import (
 	"context"
